+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [2.1.0] - 2021-01-31
 
 ### Added
 
@@ -86,7 +86,10 @@
 
 - Initial release
 
-[Unreleased]: https://github.com/paulkre/hocus/compare/v2.0.1...HEAD
+[Unreleased]: https://github.com/paulkre/hocus/compare/v2.1.0...HEAD
+[2.1.0]: https://github.com/paulkre/hocus/compare/v2.0.3...v2.1.0
+[2.0.3]: https://github.com/paulkre/hocus/compare/v2.0.2...v2.0.3
+[2.0.2]: https://github.com/paulkre/hocus/compare/v2.0.1...v2.0.2
 [2.0.1]: https://github.com/paulkre/hocus/compare/v2.0.0...v2.0.1
 [2.0.0]: https://github.com/paulkre/hocus/compare/v1.2.0...v2.0.0
 [1.2.0]: https://github.com/paulkre/hocus/compare/v1.1.1...v1.2.0